refactor(cli): split TUI into modules and serialize session control
Extract focused modules from pi-tui-runner.ts (tui-ansi colors/theme,
tui-autocomplete-layout, test terminal mock) and dedupe ANSI plus
error-handling boilerplate. Address review feedback on the control flow:

- runControl serializes model/session/permission switches so a prompt
  submitted mid-switch can't race onto the old session/model/permission
- assertSwitchable refuses /session resumes across cwd or connection,
  since autocomplete base path and model candidates are bound to the
  startup target
- keep the permission prompt visible when respondToPermission rejects,
  so y/n can retry instead of getting stuck
- showSelectPicker collapses the three pickers onto one helper

The "handles /session" test now uses a same-folder session, since
cross-folder resume is refused by design.
